+++ notepad_core/lib/NotepadConnector.dart
@@ -2,15 +2,13 @@ import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:notepad_core_platform_interface/notepad_core_platform_interface.dart';
-import 'package:convert/convert.dart';
 
 import 'Notepad.dart';
+import 'models.dart';
+import 'NotepadClient.dart';
+import 'NotepadType.dart';
 
-const SUFFIX = 'ba5e-f4ee-5ca1-eb1e5e4b1ce0';
-
-const SERV__COMMAND = '57444d01-$SUFFIX';
-const CHAR__COMMAND_REQUEST = '57444e02-$SUFFIX';
-const CHAR__COMMAND_RESPONSE = CHAR__COMMAND_REQUEST;
+typedef ConnectionChangeHandler = void Function(NotepadClient client, NotepadConnectionState state);
 
 final notepadConnector = NotepadConnector._();
 
@@ -24,7 +22,7 @@ class NotepadConnector {
   Future<dynamic> requestDevice() {
     if (!kIsWeb) throw UnimplementedError('Web platform only for now');
 
-    return NotepadCorePlatform.instance.requestDevice(optionalServices: [SERV__COMMAND]);
+    return NotepadCorePlatform.instance.requestDevice(optionalServices: NotepadClient.optionalServices);
   }
 
   void startScan() {
@@ -44,41 +42,56 @@ class NotepadConnector {
 
     return NotepadCorePlatform.instance.scanResultStream
       .map((item) => NotepadScanResult.fromMap(item))
-      .where(support);
+      .where(NotepadClient.support);
   }
 
-  void connect(scanResult) {
+  NotepadClient _notepadClient;
+  NotepadType _notepadType;
+
+  void connect(scanResult, [Uint8List authToken]) {
+    _notepadClient = NotepadClient.create(scanResult);
+    _notepadClient.setAuthToken(authToken);
+    _notepadType = NotepadType(_notepadClient);
     NotepadCorePlatform.instance.connect(scanResult);
   }
 
   void disconnect() {
+    _clean();
     NotepadCorePlatform.instance.disconnect();
   }
 
+  ConnectionChangeHandler connectionChangeHandler;
+
   Future<void> _handleMessage(NotepadCoreMessage message) async {
-    print('$_tag handleMessage $message');
-    if (message is ConnectionState) {
-      switch(message) {
-        case ConnectionState.connected:
-          await configCharacteristics();
-          await completeConnection();
-          break;
-        default:
-          print('ConnectionState ${message.value}');
-      }
+    print('$_tag _handleMessage $message');
+    if (message is NotepadConnectionState) {
+      await _handleConnectionState(message);
     }
   }
 
-  Future<void> configCharacteristics() async {
-    await NotepadCorePlatform.instance.setNotifiable(Tuple2(SERV__COMMAND, CHAR__COMMAND_REQUEST));
+  Future<void> _handleConnectionState(NotepadConnectionState connectionState) async {
+    if (connectionState == NotepadConnectionState.connected) {
+      try {
+        await _notepadType.configCharacteristics();
+        await _notepadClient.completeConnection((awaitConfirm) {
+          if (connectionChangeHandler != null) connectionChangeHandler(_notepadClient, NotepadConnectionState.awaitConfirm);
+        });
+        if (connectionChangeHandler != null) connectionChangeHandler(_notepadClient, NotepadConnectionState.connected);
+      } on AccessException {
+        _clean();
+        if (connectionChangeHandler != null) connectionChangeHandler(_notepadClient, NotepadConnectionState.disconnected);
+      }
+    } else if (connectionState == NotepadConnectionState.disconnected) {
+      _clean();
+      if (connectionChangeHandler != null) connectionChangeHandler(_notepadClient, NotepadConnectionState.disconnected);
+    } else {
+      if (connectionChangeHandler != null) connectionChangeHandler(_notepadClient, connectionState);
+    }
   }
 
-  Future<void> completeConnection() async {
-    await sendRequestAsync('Command', Tuple2(SERV__COMMAND, CHAR__COMMAND_REQUEST), Uint8List.fromList([0x01, 0x0A, 0x00, 0x00, 0x00, 0x01]));
-  }
-
-  Future<void> sendRequestAsync(String messageHead, Tuple2<String, String> serviceCharacteristic, Uint8List request) async {
-    await NotepadCorePlatform.instance.writeValue(serviceCharacteristic, request);
-    print('on${messageHead}Send: ${hex.encode(request)}');
+  // FIXME Listen to connection change
+  void _clean() {
+    _notepadClient = null;
+    _notepadType = null;
   }
 }