+++ notepad_core/lib/models.dart
@@ -1,0 +1,28 @@
+import 'dart:typed_data';
+
+class NotepadScanResult {
+  String name;
+  String deviceId;
+  Uint8List manufacturerData;
+  int rssi;
+
+  NotepadScanResult.fromMap(map)
+      : name = map['name'],
+        deviceId = map['deviceId'],
+        manufacturerData = map['manufacturerData'],
+        rssi = map['rssi'];
+
+  Map toMap() => {
+        'name': name,
+        'deviceId': deviceId,
+        'manufacturerData': manufacturerData,
+        'rssi': rssi,
+      };
+}
+
+class BatteryInfo {
+  final int percent;
+  final bool charging;
+
+  BatteryInfo(this.percent, this.charging);
+}