+++ notepad_core/lib/notepad_core.dart
@@ -1,2 +1,7 @@
-export 'Notepad.dart';
+import 'package:notepad_core_platform_interface/models.dart';
+
+export 'package:notepad_core_platform_interface/models.dart';
+
+export 'models.dart';
+export 'NotepadClient.dart';
 export 'NotepadConnector.dart';