+++ notepad_core_web/CHANGELOG.md
@@ -1,3 +1,23 @@
-## 0.0.1
+## [1.1.0] - 2019-12-20
 
-* TODO: Describe initial release.
+### NotepadCorePlatform
+
+#### Operate service/characteristic
+
+- Implement `readValue`
+- Implement `inputValueStream`
+
+## [1.0.0] - 2019-12-19
+
+### NotepadCorePlugin
+
+#### Scan/Request device
+- Implement `requestDevice`
+
+#### Connect/Disconnect device
+- Implement `connect`, `disconnect`
+- Notify `messageHandler` when connection state change
+
+#### Operate service/characteristic
+- Implement `setNotifiable`
+- Implement `writeValue`